Fix merge bug
Add a pool of 'garbage' prefixes which can be no longer used. That way
the algorithm is not dependent on the order in which addresses are
added. For instance,

merge 'abandon abandon search'
<state db: [ 'abandon abandon' ]>
merge 'abandon abandon room'
<state db: [ 'abandon abandon search', 'abandon abandon room' ]>
merge 'abandon abandon length'

previously, the shorthashes would now be
[ 'abandon abandon', 'abandon abandon search', 'abandon abandon room' ]>
now they are
[ 'abandon abandon length', 'abandon abandon search', 'abandon abandon room' ]
(since the unusable_shorthashes contains 'abandon abandon').